pyproject.toml
[project]
name = "math-ai-agent"
version = "0.0.5"
description = "Math AI Agent that uses LLM + Rubens calculator-mcp."
authors = [
    { name = "Rubens Gomes", email = "rubens.s.gomes@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12.0,<4.0.0"
dependencies = [
    "fastapi (>=0.135.1,<1.0.0)",
    "litellm (>=1.82.0,<2.0.0)",
    "fastmcp (>=3.1.0,<4.0.0)",
    "py-key-value-aio[disk] (>=0.4.4,<0.5.0)",
]

[project.urls]
Homepage = "https://github.com/rubensgomes/math-ai-agent/"
Documentation = "https://github.com/rubensgomes/math-ai-agent/README.md"
Repository = "https://github.com/rubensgomes/math-ai-agent/"

[tool.poetry]
packages = [{ include = "math_ai_agent", from = "src" }]

[build-system]
requires = ["poetry-core>=2.3.1,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
source = ["math_ai_agent"]
branch = true

[tool.coverage.report]
show_missing = true
fail_under = 90

[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)",
    "coverage (>=7.13.4,<8.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "types-pyyaml (>=6.0.12.20250915,<7.0.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
]

[tool.black]
line-length = 80

[tool.isort]
line_length = 80
profile = "black"

[tool.mypy]
mypy_path = "src"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.poe.tasks]
# Define individual cleanup tasks
_clean_cache = "poetry cache clear --all"
_clean_misc = "rm -rf  .mypy_cache dist build"
_clean_pyc = "find . -name '*.pyc' -delete"
_clean_pycache = "find . -name '__pycache__' -type d -exec rm -rf {} +"
_clean_pytest = "rm -rf .pytest_cache"
_clean_venv = "poetry env remove --all"
# Combine them into one 'clean' command
# Run command poetry run poe clean
clean = { sequence = ["_clean_cache", "_clean_misc", "_clean_pyc", "_clean_pycache", "_clean_pytest", "_clean_venv"], help = "Remove all project artifacts" }
